requirements.txt
argcomplete==2.0.0
asgiref @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_edg6nzjrx0/croots/recipe/asgiref_1661287549316/work
boto3==1.26.50
botocore==1.29.50
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
cfn-flip==1.3.0
charset-normalizer==3.0.1
click==8.1.3
Django @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_066uop7vqa/croots/recipe/django_1661292540858/work
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
durationpy==0.5
hjson==3.1.0
idna==3.4
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.1
placebo==0.9.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-slugify==7.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sqlparse @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_204sy4e5t2/croot/sqlparse_1668793448973/work
text-unidecode==1.3
toml==0.10.2
tqdm==4.64.1
troposphere==4.2.0
urllib3==1.26.14
Werkzeug==2.2.2
zappa==0.56.1
